Add files via upload
It remains to translate the names of mutations and file descriptions to the level of a new update of the modification in respect of which the translation is carried out, but basically you can enter this translation in the official mod.
Oh, yeah. There's confusion with materials like Layered Hard Material(this from russian to english without translator written) and Layered Carbide but that could be fixed if it weren't for my burnout.